# LR6/forecast/api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Weather Example API",
    "contact": {},
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "paths": {
    "/weather": {
      "get": {
        "description": "Returns current weather for given coordinates or city name. Use lat+lon or city= (Minsk, London, Tokyo, Shanghai, Warsaw). Provider: openweather (default) or google.",
        "produces": ["application/json"],
        "tags": ["weather"],
        "summary": "Get Current Weather",
        "parameters": [
          {"type": "string", "description": "Latitude (optional if city is set)", "name": "lat", "in": "query"},
          {"type": "string", "description": "Longitude (optional if city is set)", "name": "lon", "in": "query"},
          {"type": "string", "description": "City name: Minsk, London, Tokyo, Shanghai, Warsaw", "name": "city", "in": "query"},
          {"type": "string", "description": "Provider: openweather or google", "name": "provider", "in": "query", "default": "openweather"}
        ],
        "responses": {
          "200": {"description": "OK", "schema": {"$ref": "#/definitions/SuccessResponseCurrentWeather"}},
          "400": {"description": "Bad Request", "schema": {"$ref": "#/definitions/StatusResponse"}},
          "500": {"description": "Internal Server Error", "schema": {"$ref": "#/definitions/StatusResponse"}}
        }
      }
    },
    "/forecast": {
      "get": {
        "description": "Returns weather forecast for given coordinates or city name.",
        "produces": ["application/json"],
        "tags": ["weather"],
        "summary": "Get Weather Forecast",
        "parameters": [
          {"type": "string", "description": "Latitude (optional if city is set)", "name": "lat", "in": "query"},
          {"type": "string", "description": "Longitude (optional if city is set)", "name": "lon", "in": "query"},
          {"type": "string", "description": "City name: Minsk, London, Tokyo, Shanghai, Warsaw", "name": "city", "in": "query"},
          {"type": "string", "description": "Provider: openweather or google", "name": "provider", "in": "query", "default": "openweather"}
        ],
        "responses": {
          "200": {"description": "OK", "schema": {"$ref": "#/definitions/SuccessResponseForecast"}},
          "400": {"description": "Bad Request", "schema": {"$ref": "#/definitions/StatusResponse"}},
          "500": {"description": "Internal Server Error", "schema": {"$ref": "#/definitions/StatusResponse"}}
        }
      }
    },
    "/weather/batch": {
      "post": {
        "description": "Returns current temperature for several locations. Each item must have (lat+lon) or city.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["weather"],
        "summary": "Get Current Weather for Multiple Locations",
        "parameters": [
          {"in": "body", "name": "body", "required": true, "schema": {"type": "array", "items": {"$ref": "#/definitions/LocationRequest"}}},
          {"type": "string", "description": "Provider: openweather or google", "name": "provider", "in": "query", "default": "openweather"}
        ],
        "responses": {
          "200": {"description": "OK", "schema": {"$ref": "#/definitions/SuccessResponseBatchWeather"}},
          "400": {"description": "Bad Request", "schema": {"$ref": "#/definitions/StatusResponse"}},
          "500": {"description": "Internal Server Error", "schema": {"$ref": "#/definitions/StatusResponse"}}
        }
      }
    }
  },
  "definitions": {
    "StatusResponse": {
      "type": "object",
      "properties": {
        "code": {"type": "integer"},
        "message": {"type": "string"}
      }
    },
    "CurrentWeather": {
      "type": "object",
      "properties": {
        "temperature": {"type": "number"}
      }
    },
    "SuccessResponseCurrentWeather": {
      "type": "object",
      "properties": {
        "code": {"type": "integer"},
        "message": {"type": "string"},
        "data": {"$ref": "#/definitions/CurrentWeather"}
      }
    },
    "ForecastEntry": {
      "type": "object",
      "properties": {
        "date": {"type": "string"},
        "temperature": {"type": "number"}
      }
    },
    "Forecast": {
      "type": "object",
      "properties": {
        "entries": {"type": "array", "items": {"$ref": "#/definitions/ForecastEntry"}}
      }
    },
    "SuccessResponseForecast": {
      "type": "object",
      "properties": {
        "code": {"type": "integer"},
        "message": {"type": "string"},
        "data": {"$ref": "#/definitions/Forecast"}
      }
    },
    "LocationRequest": {
      "type": "object",
      "properties": {
        "lat": {"type": "string"},
        "lon": {"type": "string"},
        "city": {"type": "string"}
      }
    },
    "LocationWeather": {
      "type": "object",
      "properties": {
        "location": {"type": "string"},
        "weather": {"$ref": "#/definitions/CurrentWeather"}
      }
    },
    "SuccessResponseBatchWeather": {
      "type": "object",
      "properties": {
        "code": {"type": "integer"},
        "message": {"type": "string"},
        "data": {"type": "array", "items": {"$ref": "#/definitions/LocationWeather"}}
      }
    }
  }
}